Raise PHPStan to level 7 (#64)
* Handle false return values in File, Autoloader, XMLWriter, Drawing (PHPStan level 7)

* Fix PHPStan level 7 in PclZipAdapter, XMLReader, PasswordEncoder

* Fix PHPStan level 7 in OLERead (handle false file reads, type props and rootEntry)

* Fix PHPStan level 7 in test helpers (AbstractZipAdapter, TestHelperZip, XmlDocument)

* Bump PHPStan analysis to level 7

* Fix PHPStan level 7 across the whole PHP matrix (7.1 to 8.5)

* Fix PHPStan level 7 on PHP 7.2-7.4 (substr, mb_convert_encoding, spl_autoload_functions)

* Use string casts instead of guards to keep test coverage stable

+++ src/Common/Drawing.php
@@ -140,9 +140,9 @@ class Drawing
             return null;
         }
 
-        $colorR = hexdec($colorR);
-        $colorG = hexdec($colorG);
-        $colorB = hexdec($colorB);
+        $colorR = (int) hexdec($colorR);
+        $colorG = (int) hexdec($colorG);
+        $colorB = (int) hexdec($colorB);
 
         return [$colorR, $colorG, $colorB];
     }
